requirements.txt
fastapi
uvicorn[standard]
praw
groq
python-dotenv
